fix: move test to another file
